Improve package structure and visibility
