+++ testutil-server/src/main/java/io/spine/testing/server/blackbox/BlackBoxBoundedContext.java
@@ -61,12 +61,10 @@ import io.spine.system.server.event.CommandErrored;
 import io.spine.testing.client.TestActorRequestFactory;
 import io.spine.testing.server.CommandSubject;
 import io.spine.testing.server.EventSubject;
-import io.spine.testing.server.SubscriptionActivator;
 import io.spine.testing.server.SubscriptionObserver;
 import io.spine.testing.server.VerifyingCounter;
-import io.spine.testing.server.blackbox.verify.query.QueryResultSubject;
-import io.spine.testing.server.blackbox.verify.subscription.ToProtoSubjects;
 import io.spine.testing.server.entity.EntitySubject;
+import io.spine.testing.server.query.QueryResultSubject;
 import io.spine.time.ZoneId;
 import io.spine.time.ZoneOffset;
 import io.spine.type.TypeName;
